make cache_status_new the default, rename previous one to _old
